requirements.txt
ipython
jupyter
chardet
pyyaml
pytest
pytest-cov